Updated unequal length tests in hamming (#168)
* Updated unequal length tests in hamming

* Updated hamming tests for error conditions.

* ran indent.sh

* Updated hamming example.c per code review comments

* Removed strlen from within loop on hamming example.c

* Made hamming return type int to account for '-1' error case.

+++ exercises/hamming/src/example.h
@@ -1,6 +1,6 @@
 #ifndef HAMMING_H
 #define HAMMING_H
 
-size_t compute(const char *lhs, const char *rhs);
+int compute(const char *lhs, const char *rhs);
 
 #endif
